MinSize is now int64

+++ middleware/gzip/middleware.go
@@ -69,7 +69,7 @@ func (grw *gzipResponseWriter) Write(b []byte) (int, error) {
 
 	// Check if length is above minimum,
 	// if not save to buffer.
-	size := len(b) + grw.buf.Len()
+	size := int64(len(b)) + int64(grw.buf.Len())
 	if size < grw.o.minSize {
 		return grw.buf.Write(b)
 	}
@@ -110,7 +110,7 @@ type Option func(*options) error
 // options contains final options
 type options struct {
 	ignoreRange  bool
-	minSize      int
+	minSize      int64
 	contentTypes []string
 	statusCodes  map[int]struct{}
 }
@@ -220,7 +220,7 @@ func OnlyStatusCodes(codes ...int) Option {
 }
 
 // MinSize will set a minimum size for compression.
-func MinSize(n int) Option {
+func MinSize(n int64) Option {
 	return func(c *options) error {
 		if n <= 0 {
 			c.minSize = 0
